[config]: set target
