+++ JUN-tong/JUN-tong/CityBusTimeCell.swift
@@ -30,7 +30,9 @@ class CityBusTimeCell: UITableViewCell {
             UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound, .badge],
                                                                     completionHandler: {didAllow, error in
                                                                         if didAllow {
-                                                                            print("allow")
+                                                                            DispatchQueue.global(qos: .userInteractive).async {
+                                                                                self.alarmTimePicekr()
+                                                                            }
                                                                         } else {
                                                                             let alert = UIAlertController(title: "알람 설정 필요",
                                                                                                           message: "알람을 받기 위해서는 알람 설정이 필요합니다.\n알람 설정은 '설정'탭에서 확인할 수 있습니다.", preferredStyle: .alert)
@@ -41,21 +43,22 @@ class CityBusTimeCell: UITableViewCell {
             })
         }
     }
+    
+    func alarmTimePicekr() {
+        DispatchQueue.main.async {
+            let timePicker = UIDatePicker()
+            timePicker.datePickerMode = UIDatePickerMode.countDownTimer
+            if let parentViewSize = self.parentViewController?.view.frame.size {
+                timePicker.frame.size.height = parentViewSize.height
+                timePicker.frame.size.width = parentViewSize.width
+            }
 
-    private func alarmResister() {
-
-    }
-}
-
-extension UIView {
-    var parentViewController: UIViewController? {
-        var parentResponder: UIResponder? = self
-        while parentResponder != nil {
-            parentResponder = parentResponder!.next
-            if let viewController = parentResponder as? UIViewController {
-                return viewController
+            if let parentView = self.parentViewController {
+                timePicker.frame.origin = CGPoint(x: 0,
+                                                  y: 0)
+                timePicker.backgroundColor = UIColor.white
+                parentView.view.addSubview(timePicker)
             }
         }
-        return nil
     }
 }
