+++ JUN-tong/JUN-tong/MainViewController.swift
@@ -342,7 +342,7 @@ extension MainViewController: UITableViewDelegate, UITableViewDataSource {
     func numberOfSections(in tableView: UITableView) -> Int {
         return 2
     }
-
+    
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
         if section == 0 && favoriteBusList.isEmpty {
             return 0
