+++ JUN-tong/JUN-tong/UIView+extension.swift
@@ -6,4 +6,17 @@
 //  Copyright © 2018년 Seong ho Hong. All rights reserved.
 //
 
-import Foundation
+import UIKit
+
+extension UIView {
+    var parentViewController: UIViewController? {
+        var parentResponder: UIResponder? = self
+        while parentResponder != nil {
+            parentResponder = parentResponder!.next
+            if let viewController = parentResponder as? UIViewController {
+                return viewController
+            }
+        }
+        return nil
+    }
+}
